Send events thorough a channel to let users build histograms
Users of the library might want to build histograms of things like how
long goroutines are waiting for teh pool. Rather than building in some
kind of histograming solution (with what bins??), provide the events to
the user over a channel so that if they choose, they can consume the
channel and handle the timing data however they want.