[OUT-5021] Ensure that MasterConns stat is a counter
I believe the MasterConns stat is supposed to be a counter that is never
decrementing. However, it actually does get decremented in the pool
shrinker.

Having it be an always-incrementing counter means you can use your
monitoring to do things like measure the rate at which connections are
opened.

If it was supposed to be decremented on connection close, it should be
done in kill() where it would also happen on connection failure or
abnormal connection close.