[OUT-5021] Add metrics for shrunk and failed connections
Since I took away the change to the MasterConns and SlaveConns metrics
in the pool shrinker, I replaced it with a metric to capture when the
pool shrinking happens.

I also added a metric incrmeneted when a connection is abended. Frequent
opening and closing of connections is a sign of a serious problem with
Mongo and it's something I'd like to keep tabs on.